Move http2 module to core
